requirements.txt
numpy~=1.24.3
pandas~=1.5.3
autograd~=1.6.2
tensorflow~=2.10.0
tensorflow-probability~=0.18.0
matplotlib~=3.7.3
scikit-survival~=0.21.0
scikit-learn~=1.2.2
notebook~=7.0.4
seaborn~=0.12.2
wandb~=0.15.12
pyarrow~=13.0.0
lifelines~=0.26.4
torch~=1.13.1
torchvision~=0.14.1
pycox~=0.2.2
scipy~=1.10.1
scikit-multilearn~=0.2.0
sklearn_pandas~=2.2.0
tqdm~=4.66.2
rpy2==3.5.1
setuptools~=52.0.0
tzlocal~=5.1
auton-survival~=0.1.0
tf-models-official~=2.10.0